planets with different radius and adjusted gravity, planets are only rendered when on screen.
