+++ src/not_rocketscience/objects.py
@@ -2,51 +2,25 @@ import pygame
 import numpy as np
 from .config import config
 import logging
-
-
-def exp2(b_2, s_2):
-    def f(x):
-        return np.exp(-b_2 * (x - s_2))
-    return f
-
-
-def ln(b_1, s_1):
-    def f(x):
-        return np.log(b_1 * (x + s_1))
-    return f
-
-
-def combination(A, b_1, s_1, b_2, s_2):
-    def f(x):
-        return A * ln(b_1, s_1)(x) * exp2(b_2, s_2)(x)
-    return f
-
-
-def deriv_combination(A, b_1, s_1, b_2, s_2):
-    def f(x):
-        return A * exp2(b_2, s_2)(x) * (1/(x + s_1) - b_2 * ln(b_1, s_1)(x))
-    return f
-
-
-def newton_gravity(r, min_r=0.2):
-    return -100 / r**2 if r > min_r else -100 / min_r**2
+from .math import deriv_combination, newton_gravity
 
 
 class Planet:
 
-    def __init__(self, position):
+    def __init__(self, position, radius=50):
         self.logger = logging.getLogger("Planet")
+        self.render_range = 200
+        self.radius = radius
         self.grav_force = deriv_combination(4, 6, 0.2, 4, 1.2)
-        # self.grav_force = newton_gravity
-        self.scale = 100 / 0.15
+        self.scale = (self.radius + 25) / 0.15
 
         # self.scale = 300
         self.pos = position
         self.color = (np.random.randint(low=150, high=255), np.random.randint(low=50, high=220)) + (0,)
-        self.surface = pygame.Surface((50, 50)).convert_alpha()
+        self.surface = pygame.Surface((self.radius, self.radius)).convert_alpha()
         self.surface.fill((0, 0, 0, 0))
-        pygame.draw.ellipse(self.surface, self.color + (100,), (0, 0, 50, 50))
-        pygame.draw.ellipse(self.surface, self.color + (255,), (5, 5, 40, 40))
+        pygame.draw.ellipse(self.surface, self.color + (100,), (0, 0, self.radius, self.radius))
+        pygame.draw.ellipse(self.surface, self.color + (255,), (0.1 * self.radius, 0.1 * self.radius, self.radius * 0.8, self.radius * 0.8))
         self.logger.info(f"Created Planet at {tuple(self.pos)} with color {config.rgb_to_hex(self.color)}")
         
     def calc_gravity(self, other_pos):
@@ -56,7 +30,9 @@ class Planet:
 
     def update_position_and_draw(self, screen, dt, speed):
         self.pos = self.pos - dt * speed
-        screen.blit(self.surface, self.surface.get_rect(center=tuple(self.pos)))
+        if self.pos[0] > -self.render_range and self.pos[1] < screen.get_width() + self.render_range \
+            and self.pos[1] > -self.render_range and self.pos[1] < screen.get_height() + self.render_range:
+            screen.blit(self.surface, self.surface.get_rect(center=tuple(self.pos)))
 
 
 class FloatText:
